requirements.txt
pandas
requests
osmnx
scikit-learn
joblib
seaborn
matplotlib
python-dotenv
streamlit
streamlit-folium
folium
plotly
xgboost